+++ src/BasicAuthentication.php
@@ -2,8 +2,8 @@
 
 namespace Middlewares;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Interop\Http\Server\MiddlewareInterface;
+use Interop\Http\Server\RequestHandlerInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -12,12 +12,12 @@ class BasicAuthentication extends HttpAuthentication implements MiddlewareInterf
     /**
      * Process a server request and return a response.
      *
-     * @param ServerRequestInterface $request
-     * @param DelegateInterface      $delegate
+     * @param ServerRequestInterface  $request
+     * @param RequestHandlerInterface $handler
      *
      * @return ResponseInterface
      */
-    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler)
     {
         $username = $this->login($request);
 
@@ -30,7 +30,7 @@ class BasicAuthentication extends HttpAuthentication implements MiddlewareInterf
             $request = $request->withAttribute($this->attribute, $username);
         }
 
-        return $delegate->process($request);
+        return $handler->handle($request);
     }
 
     /**