+++ src/DigestAuthentication.php
@@ -2,8 +2,8 @@
 
 namespace Middlewares;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Interop\Http\Server\MiddlewareInterface;
+use Interop\Http\Server\RequestHandlerInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -31,12 +31,12 @@ class DigestAuthentication extends HttpAuthentication implements MiddlewareInter
     /**
      * Process a server request and return a response.
      *
-     * @param ServerRequestInterface $request
-     * @param DelegateInterface      $delegate
+     * @param ServerRequestInterface  $request
+     * @param RequestHandlerInterface $handler
      *
      * @return ResponseInterface
      */
-    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler)
     {
         $username = $this->login($request);
 
@@ -56,7 +56,7 @@ class DigestAuthentication extends HttpAuthentication implements MiddlewareInter
             $request = $request->withAttribute($this->attribute, $username);
         }
 
-        return $delegate->process($request);
+        return $handler->handle($request);
     }
 
     /**